Abstract Graph Reporter Class

Defines the Abstract Class for all PackageGraphReporters defined in pkgnet. The class is not meant to be instantiated, but inherited from and its methods overloaded such that each Metric implements certain functionality.

AbstractGraphReporter

Format

An object of class R6ClassGenerator of length 24.

Public Methods

calculate_network_measures()

extract network features on a node and network level

Returns: A list containing:

networkMeasures: a list of network measures.

nodeMeasures: the nodes data.table with additional columns of node measures.

plot_network()

Creates a network visualization of extracted package graph.

Args:

...: ...

Returns:

A `visNetwork` object

Public Members

edges

A data.table from SOURCE to TARGET nodes describing the connections

nodes

A data.table with node as an identifier, and augmenting information about each node

pkgGraph

An igraph object describing the package graph

networkMeasures

A list of network measures calculated by calculate_network_features

layoutType

Character string indicating currently active graph layout

graphViz

visNetwork object of package graph

Active Bindings

pkgGraph()

Returns the graph object

networkMeasures()

Returns a table of network measures, one row per node

graphViz()

Returns ths graph visualization object

orphanNodes()

Returns the list of orphan nodes

layoutType(value)

If no value given, the current layout type for the graph visualization is returned. If a vaild layput type is given, this fucntion will update the layoutType field.

orphanNodeClusteringThreshold(value)

If no value given, the current orphan node clustering threshold is returned. If a valid orphan node clustering threshold is given, this function will update the orphan node clustering threshold.

See also

Other AbstractReporters: AbstractPackageReporter