Abstract Package Reporter Class

Defines the Abstract Class for all PackageReporters defined in pkgnet. The class is not meant to be instantiated, but inherited from and its methods overloaded such that each Metric implements certain functionality.

AbstractPackageReporter

Format

An object of class R6ClassGenerator of length 24.

Public

get_package()

Returns a string with the name of the package used to construct this object

set_package(packageName)

Set the package that all operations in the object are done for.

Args:

packageName: a string with the name of the package you are analyzing.

packagePath: directory path to source code of package

get_report()

Returns a particular reporter's report on the package

get_report_markdown_path()

Returns the path to the markdown report associated with this reporter

get_summary_view()

Returns a particular reporters summary report on the package for use in a high level view

calculate_package_metrics()

Calculates all the relevant metrics for a set package and updates the reporter

get_raw_data()

Returns a particular reporter's raw data.

Abstract Methods

getPackageMetrics

Accepts a package name as a string and calculates the PackageMetrics required for that reporter

generateReport

Generates a a child report for an overall knit package report

getSummaryView

Returns a summarized view of the package metrics as a data.table

getRawData

Returns the internal data of the packageReporter as a list

See also

Other AbstractReporters: AbstractGraphReporter