pkgnet Report as Vignette

Create pkgnet package report as an R Markdown vignette. This vignette can be rendered into a standard HTML vignette with the knitr::rmarkdown vignette engine into HTML vignettes upon package building. It is also compatible with #' pkgdown sites. See the vignette "Publishing Your pkgnet Package Report" for details about how to use this function, as well as our example for pkgnet.

CreatePackageVignette(
  pkg = ".",
  pkg_reporters = list(DependencyReporter$new(), FunctionReporter$new()),
  vignette_path = file.path(pkg, "vignettes", "pkgnet-report.Rmd")
)

Arguments

pkg

(string) path to root directory of package of interest

pkg_reporters

(list) a list of initialized package reporters

vignette_path

(string) The location of a file to store the output vignette file at. Must be an .Rmd file. By default, this will be '<pkg>/vignettes/pkgnet-report.Rmd' relative to the input to pkg