Recursive Package Dependency Reporter

This reporter looks at the recursive network of its dependencies on other packages. This allows a developer to understand how individual dependencies might lead to a much larger set of dependencies, potentially informing decisions on including or removing them.

See also

Super classes

pkgnet::AbstractPackageReporter -> pkgnet::AbstractGraphReporter -> DependencyReporter

Active bindings

report_markdown_path

(character string) path to R Markdown template for this reporter. Read-only.

Methods

Public methods

DependencyReporter$new()

DependencyReporter$clone()

Inherited methods

pkgnet::AbstractPackageReporter$set_package()

pkgnet::AbstractGraphReporter$calculate_default_measures()

pkgnet::AbstractGraphReporter$get_summary_view()

Method new()

Initialize an instance of the reporter.

Usage

DependencyReporter$new(
  dep_types = c("Imports", "Depends", "LinkingTo"),
  installed = TRUE
)

Arguments

dep_types

(character vector) The sections within the DESCRIPTION file to be counted as dependencies. By default, c("Imports", "Depends", "LinkingTo") is chosen.

installed

(logical) If TRUE, consider only installed packages when building dependency network.

Returns

Self, invisibly.

Examples

\donttest{

# Instantiate an object
reporter <- DependencyReporter$new()

# Seed it with a package
reporter$set_package("ggplot2")

}

Method clone()

The objects of this class are cloneable with this method.

Usage

DependencyReporter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `DependencyReporter$new`
## ------------------------------------------------

# \donttest{

# Instantiate an object
reporter <- DependencyReporter$new()

# Seed it with a package
reporter$set_package("ggplot2")

# }