Directed Graph Network Model

R6 class defining a directed graph model for representing a network, including methods to calculate various measures from graph theory. The igraph package is used as a backend for calculations.

This class isn't intended to be initialized directly; instead, network reporter objects will initialize it as its pkg_graph field. If you have a network reporter named reporter, then you access this object's public interface through pkg_graph---for example,

reporter$pkg_graph$node_measures('hubScore')

See also

Super class

pkgnet::AbstractGraph -> DirectedGraph

Active bindings

default_node_measures

character vector of default node measures. See Node Measures section in DirectedGraphMeasures for details about each measure. Read-only.

default_graph_measures

character vector of default graph measures. See Graph Measures section in DirectedGraphMeasures for details about each measure. Read-only.

Methods

Public methods

DirectedGraph$clone()

Inherited methods

pkgnet::AbstractGraph$graph_measures()

pkgnet::AbstractGraph$initialize()

pkgnet::AbstractGraph$node_measures()

pkgnet::AbstractGraph$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

DirectedGraph$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.