Plot Network Structure

Plot the network structure of an R package.

PlotNetwork(edges, nodes, colorFieldName = NULL)

Arguments

edges
a data.table with SOURCE and TARGET nodes

nodes
a data.table with node and augmenting metadata

colorFieldName
(character) The name of column to use to color the nodes. This can be any column in the "nodes" table inside the object produced by ExtractFunctionNetwork. Default is outDegree. If you estimated test coverage when creating the network representation, try setting this field to "test_coverage"!

Value

A plotly object (local, not on plot.ly)

Examples

# NOT RUN {
nw <- ExtractFunctionNetwork("lubridate")
nw$nodes[,test := runif(.N)]
PlotNetwork(nw, colorFieldName = "test")
# }