Abstract Package Reporter

pkgnet defines several package reporter R6 classes that analyze some particular aspect of a package. These reporters share common functionality and interfaces defined by a base reporter class AbstractPackageReporter.

Active bindings

pkg_name

(character string) name of set package. Read-only.

report_markdown_path

(character string) path to R Markdown template for this reporter. Read-only.

Methods

Public methods

AbstractPackageReporter$set_package()

AbstractPackageReporter$get_summary_view()

AbstractPackageReporter$clone()

Method set_package()

Set the package that the reporter will analyze. This can only be done once for a given instance of a reporter. Instantiate a new copy of the reporter if you need to analyze a different package.

Usage

AbstractPackageReporter$set_package(pkg_name, pkg_path = NULL)

Arguments

pkg_name

(character string) name of package

pkg_path

(character string) optional directory path to source code of the package. It is used for calculating test coverage. It can be an absolute or relative path.

Returns

Self, invisibly.

Method get_summary_view()

Returns an htmlwidget object that summarizes the analysis of the reporter. Used when creating a package report.

Usage

AbstractPackageReporter$get_summary_view()

Returns

Self, invisibly.

Method clone()

The objects of this class are cloneable with this method.

Usage

AbstractPackageReporter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.