Recursive Package Dependency Reporter

This reporter looks at the recursive network of its dependencies on other packages. This allows a developer to understand how individual dependencies might lead to a much larger set of dependencies, potentially informing decisions on including or removing them.

Class Constructor

DependencyReporter$new()

Initialize an instance of the reporter.

Returns:

Instantiated reporter object. Note that this reporter object isn't useful yet until you use the set_package method to set a package.

Public Methods

set_package(pkg_name, pkg_path = NULL)

Set the package that the reporter will analyze. This can only be done once for a given instance of a reporter. Instantiate a new copy of the reporter if you need to analyze a different package.

Args:

pkg_name: character string, name of package

pkg_path: character string, optional directory path to source code of the package. It is used for calculating test coverage. It can be an absolute or relative path.

Returns:

Self, invisibly.

get_summary_view()

Returns an htmlwidget object that summarizes the analysis of the reporter. Used when creating a package report.

Returns:

htmlwidget object

calculate_default_measures()

Calculates the default node and network measures for this reporter.

Returns:

Self, invisibly.

Public Fields

pkg_name

character string, name of set package. Read-only.

report_markdown_path

character string, path to R Markdown template for this reporter. Read-only.

nodes

a data.table, containing information about the nodes of the network the reporter is analyzing. The node column acts the identifier. Read-only.

edges

a data.table, containing information about the edge connections of the network the reporter is analyzing. Each row is one edge, and the columns SOURCE and TARGET specify the node identifiers. Read-only.

network_measures

a list, containing any measures of the network calculated by the reporter. Read-only.

pkg_graph

a graph model object. See DirectedGraph for additional documentation. Read-only.

graph_viz

a graph visualization object. A visNetwork::visNetwork object. Read-only.

layout_type

a character string, the current layout type for the graph visualization. Can be assigned a new valid layout type value. Use use grep("^layout_\\S", getNamespaceExports("igraph"), value = TRUE) to see valid options.

Special Methods

clone(deep = FALSE)

Method for copying an object. See Advanced R for the intricacies of R6 reference semantics.

Args:

deep: logical. Whether to recursively clone nested R6 objects.

Returns:

Cloned object of this class.

See also

Examples

# \donttest{

# Instantiate an object
reporter <- DependencyReporter$new()

# Seed it with a package
reporter$set_package("ggplot2")
#> FATAL [2021-12-23 12:50:38] pkgnet could not find an installed package named 'ggplot2'. Please install the package first.
#> Error in log_fatal(msg): pkgnet could not find an installed package named 'ggplot2'. Please install the package first.

# }