Measures for Directed Graph Class

Descriptions for all available node and graph measures for networks modeled by DirectedGraph.

Node Measures

outDegree

: outdegree, the number of outward edges (tail ends). Calculated by igraph::degree. [Wikipedia]

inDegree

: indegree, number of inward edges (head ends). Calculated by igraph::degree. [Wikipedia]

outCloseness

: closeness centrality (out), a measure of path lengths to other nodes along edge directions. Calculated by igraph::closeness. [Wikipedia]

inCloseness

: closeness centrality (in), a measure of path lengths to other nodes in reverse of edge directions. Calculated by igraph::closeness. [Wikipedia]

numRecursiveDeps

: number recursive dependencies, i.e., count of all nodes reachable by following edges out from this node. Calculated by igraph::neighborhood.size. [Wikipedia]

numRecursiveRevDeps

: number of recursive reverse dependencies (dependents), i.e., count all nodes reachable by following edges into this node in reverse direction. Calculated by igraph::neighborhood.size. [Wikipedia]

betweenness

: betweenness centrality, a measure of the number of shortest paths in graph passing through this node Calculated by igraph::betweenness. [Wikipedia]

pageRank

: Google PageRank. Calculated by igraph::page_rank. [Wikipedia]

hubScore

: hub score from Hyperlink-Induced Topic Search (HITS) algorithm. Calculated by igraph::hub_score. [Wikipedia]

authorityScore

: authority score from Hyperlink-Induced Topic Search (HITS) algorithm. Calculated by igraph::authority_score. [Wikipedia]

Graph Measures

graphOutDegree

: graph freeman centralization for outdegree. A measure of the most central node by outdegree in relation to all other nodes. Calculated by igraph::centralize. [Wikipedia]

graphInDegree

: graph Freeman centralization for indegree. A measure of the most central node by indegree in relation to all other nodes. Calculated by igraph::centralize. [Wikipedia]

graphOutClosness

: graph Freeman centralization for out-closeness. A measure of the most central node by out-closeness in relation to all other nodes. Calculated by igraph::centralize. [Wikipedia]

graphInCloseness

: graph Freeman centralization for outdegree. A measure of the most central node by outdegree in relation to all other nodes. Calculated by igraph::centralize. [Wikipedia]

graphBetweennness

: graph Freeman centralization for betweenness A measure of the most central node by betweenness in relation to all other nodes. Calculated by igraph::centralize. [Wikipedia]