Package Summary Reporter Class
Source: R/PackageSummaryReporter.R

Defines a concrete implementation of AbstractPackageReporter for a high level overview of a particular package. It will summarize things like lines of code, whether it's on CRAN, etc.

SummaryReporter

Format

An object of class R6ClassGenerator of length 24.

Public Methods

set_package(pkg_name, pkg_path = NULL)

Set the package that all operations in the object are done for.

Args:

pkg_name: A string with the name of the package you are analyzing.

pkg_path: Optional directory path to source code of the package. It is used for calculating test coverage. It can be an absolute or relative path.

get_summary_view()

Returns a particular reporters summary report on the package for use in a high level view

See also